Split Ares into smaller pieces, hopfully to break up the dependencies.
