+++ src/ares/AresHandler.ts
@@ -10,17 +10,16 @@ import {ResourceType} from '../ResourceType';
 import {SpaceBonus} from '../SpaceBonus';
 import {TileType} from '../TileType';
 import {ITile} from '../ITile';
-import {IAresData, IHazardConstraint, IMilestoneCount} from './IAresData';
+import {IAresData, IMilestoneCount} from './IAresData';
 import {IAdjacencyCost} from './IAdjacencyCost';
-import {ARES_MILESTONES} from '../milestones/Milestones';
-import {ARES_AWARDS} from '../awards/Awards';
 import {Multiset} from '../utils/Multiset';
 import {Phase} from '../Phase';
 import {DeferredAction} from '../deferredActions/DeferredAction';
 import {AddResourcesToCard} from '../deferredActions/AddResourcesToCard';
 import {SelectHowToPayDeferred} from '../deferredActions/SelectHowToPayDeferred';
 import {SelectProductionToLoseDeferred} from '../deferredActions/SelectProductionToLoseDeferred';
-import {IDrawnMilestonesAndAwards} from '../MilestoneAwardSelector';
+import {_AresHazardPlacement} from './AresHazards';
+import {bonusAsString, tileTypeAsString} from './_StringGen';
 
 export const OCEAN_UPGRADE_TILES = [TileType.OCEAN_CITY, TileType.OCEAN_FARM, TileType.OCEAN_SANCTUARY];
 export const HAZARD_TILES = [TileType.DUST_STORM_MILD, TileType.DUST_STORM_SEVERE, TileType.EROSION_MILD, TileType.EROSION_SEVERE];
@@ -33,27 +32,6 @@ export enum HazardSeverity {
 
 export class AresHandler {
   private constructor() {}
-
-  public static initialData(active: boolean, includeHazards: boolean, players: Player[]): IAresData {
-    return {
-      active: active,
-      includeHazards: includeHazards,
-      hazardData: {
-        erosionOceanCount: {threshold: 3, available: true}, // oceans: add erosion tiles
-        removeDustStormsOceanCount: {threshold: 6, available: true}, // oceans: remove dust storms
-        severeErosionTemperature: {threshold: -4, available: true}, // temperatore: severe erosion
-        severeDustStormOxygen: {threshold: 5, available: true}, // oxygen: severe dust storms
-      },
-      milestoneResults: players.map((p) => {
-        return {id: p.id, count: 0};
-      }),
-    };
-  }
-
-  public static setupMilestonesAwards(drawnMilestonesAndAwards: IDrawnMilestonesAndAwards) {
-    drawnMilestonesAndAwards.milestones.push(...ARES_MILESTONES);
-    drawnMilestonesAndAwards.awards.push(...ARES_AWARDS);
-  }
 
   public static ifAres(game: Game, cb: (aresData: IAresData) => void) {
     if (game.gameOptions.aresExtension) {
@@ -225,23 +203,6 @@ export class AresHandler {
     return count;
   }
 
-  public static setupHazards(game: Game, playerCount: number) {
-    // The number of dust storms depends on the player count.
-    // I made up that the solo player has 3 dust storms. The rules
-    // don't take solo into account, nor if you played with more than
-    // five players.
-    if (playerCount >= 5) {
-      randomlyPlaceHazard(game, TileType.DUST_STORM_MILD, 1);
-    } else if (playerCount === 4) {
-      randomlyPlaceHazard(game, TileType.DUST_STORM_MILD, 1);
-      randomlyPlaceHazard(game, TileType.DUST_STORM_MILD, -1);
-    } else if (playerCount <= 3) {
-      randomlyPlaceHazard(game, TileType.DUST_STORM_MILD, 1);
-      randomlyPlaceHazard(game, TileType.DUST_STORM_MILD, 1);
-      randomlyPlaceHazard(game, TileType.DUST_STORM_MILD, -1);
-    }
-  }
-
   public static hasHazardTile(space: ISpace): boolean {
     return AresHandler.hazardSeverity(space) !== HazardSeverity.NONE;
   }
@@ -352,26 +313,15 @@ export class AresHandler {
   }
 
   public static onTemperatureChange(game: Game, aresData: IAresData) {
-    // This will have no effect if the erosions don't exist, but that's OK --
-    // the check for placing erosions will take this into account.
-    testConstraint(
-      aresData.hazardData.severeErosionTemperature,
-      game.getTemperature(),
-      () => {
-        makeSevere(game, TileType.EROSION_MILD, TileType.EROSION_SEVERE);
-      },
-    );
+    _AresHazardPlacement.onTemperatureChange(game, aresData);
   }
 
   public static onOceanPlaced(game: Game, aresData: IAresData, player: Player) {
-    testToPlaceErosionTiles(game, aresData, player);
-    testToRemoveDustStorms(game, aresData, player);
+    _AresHazardPlacement.onOceanPlaced(game, aresData, player);
   }
 
   public static onOxygenChange(game: Game, aresData: IAresData) {
-    testConstraint(aresData.hazardData.severeDustStormOxygen, game.getOxygenLevel(), () => {
-      makeSevere(game, TileType.DUST_STORM_MILD, TileType.DUST_STORM_SEVERE);
-    });
+    _AresHazardPlacement.onOxygenChange(game, aresData);
   }
 
   public static grantBonusForRemovingHazard(game: Game, player: Player, initialTileType: TileType | undefined) {
@@ -396,137 +346,5 @@ export class AresHandler {
     player.increaseTerraformRatingSteps(steps, game);
     game.log('${0}\'s TR increases ${1} step(s) for removing ${2}', (b) => b.player(player).number(steps).string(tileTypeAsString(initialTileType)));
   }
-
-
-  public static putHazardAt(space: ISpace, tileType: TileType) {
-    space.tile = {tileType: tileType, protectedHazard: false};
-  }
 }
 
-function randomlyPlaceHazard(game: Game, tileType: TileType, direction: 1 | -1) {
-  const space = game.getSpaceByOffset(direction, 'hazard');
-  AresHandler.putHazardAt(space, tileType);
-  return space;
-}
-
-
-function makeSevere(game: Game, from: TileType, to: TileType) {
-  game.board.spaces
-    .filter((s) => s.tile?.tileType === from)
-    .forEach((s) => {
-      AresHandler.putHazardAt(s, to);
-    });
-
-  game.log('${0} have upgraded to ${1}', (b) => b.string(tileTypeAsString(from)).string(tileTypeAsString(to)));
-}
-
-function testConstraint(constraint: IHazardConstraint, testValue: number, cb: () => void) {
-  if (!constraint.available) {
-    return;
-  }
-  if (testValue >= constraint.threshold) {
-    cb();
-    constraint.available = false;
-  }
-}
-
-function testToRemoveDustStorms(game: Game, aresData: IAresData, player: Player) {
-  testConstraint(
-    aresData.hazardData.removeDustStormsOceanCount,
-    game.board.getOceansOnBoard(),
-    () => {
-      game.board.spaces.forEach((space) => {
-        if (space.tile?.tileType === TileType.DUST_STORM_MILD || space.tile?.tileType === TileType.DUST_STORM_SEVERE) {
-          if (space.tile.protectedHazard !== true) {
-            space.tile = undefined;
-          }
-        }
-      });
-
-      if (game.phase !== Phase.SOLAR) {
-        player.increaseTerraformRating(game);
-        game.log('${0}\'s TR increases 1 step for eliminating dust storms.', (b) => b.player(player));
-      }
-    },
-  );
-}
-
-function testToPlaceErosionTiles(game: Game, aresData: IAresData, player: Player) {
-  testConstraint(
-    aresData.hazardData.erosionOceanCount,
-    game.board.getOceansOnBoard(),
-    () => {
-      let type = TileType.EROSION_MILD;
-      if (aresData.hazardData.severeErosionTemperature.available !== true) {
-        type = TileType.EROSION_SEVERE;
-      }
-
-      const space1 = randomlyPlaceHazard(game, type, 1);
-      const space2 = randomlyPlaceHazard(game, type, -1);
-      [space1, space2].forEach((space) => {
-        LogHelper.logTilePlacement(game, player, space, type);
-      });
-    },
-  );
-}
-
-// TODO(kberg): convert to a log message type
-const tileTypeToStringMap: Map<TileType, string> = new Map([
-  [TileType.GREENERY, 'greenery'],
-  [TileType.OCEAN, 'ocean'],
-  [TileType.CITY, 'city'],
-
-  [TileType.CAPITAL, CardName.CAPITAL],
-  [TileType.COMMERCIAL_DISTRICT, CardName.COMMERCIAL_DISTRICT],
-  [TileType.ECOLOGICAL_ZONE, CardName.ECOLOGICAL_ZONE],
-  [TileType.INDUSTRIAL_CENTER, CardName.INDUSTRIAL_CENTER],
-  [TileType.LAVA_FLOWS, CardName.LAVA_FLOWS],
-  [TileType.MINING_AREA, CardName.MINING_AREA],
-  [TileType.MINING_RIGHTS, CardName.MINING_RIGHTS],
-  [TileType.MOHOLE_AREA, CardName.MOHOLE_AREA],
-  [TileType.NATURAL_PRESERVE, CardName.NATURAL_PRESERVE],
-  [TileType.NUCLEAR_ZONE, CardName.NUCLEAR_ZONE],
-  [TileType.RESTRICTED_AREA, CardName.RESTRICTED_AREA],
-
-  // DEIMOS_DOWN, "",
-  // GREAT_DAM, "",
-  // MAGNETIC_FIELD_GENERATORS, "",
-
-  [TileType.BIOFERTILIZER_FACILITY, CardName.BIOFERTILIZER_FACILITY],
-  [TileType.METALLIC_ASTEROID, CardName.METALLIC_ASTEROID],
-  [TileType.SOLAR_FARM, CardName.SOLAR_FARM],
-  [TileType.OCEAN_CITY, CardName.OCEAN_CITY],
-  [TileType.OCEAN_FARM, CardName.OCEAN_FARM],
-  [TileType.OCEAN_SANCTUARY, CardName.OCEAN_SANCTUARY],
-  [TileType.DUST_STORM_MILD, 'Mild Dust Storm'],
-  [TileType.DUST_STORM_SEVERE, 'Severe Dust Storm'],
-  [TileType.EROSION_MILD, 'Mild Erosion'],
-  [TileType.EROSION_SEVERE, 'Severe Erosion'],
-  [TileType.MINING_STEEL_BONUS, 'Mining (Steel)'],
-  [TileType.MINING_TITANIUM_BONUS, 'Mining (Titanium)'],
-]);
-
-function tileTypeAsString(tileType: TileType | undefined): string {
-  if (tileType === undefined) {
-    return 'undefined';
-  };
-  return tileTypeToStringMap.get(tileType)|| 'special';
-}
-
-
-const bonusToStringMap: Map<SpaceBonus, string> = new Map([
-  [SpaceBonus.TITANIUM, 'Titanium'],
-  [SpaceBonus.STEEL, 'Steel'],
-  [SpaceBonus.PLANT, 'Plant'],
-  [SpaceBonus.DRAW_CARD, 'Card'],
-  [SpaceBonus.HEAT, 'Heat'],
-  [SpaceBonus.OCEAN, 'Ocean'],
-  [SpaceBonus.MEGACREDITS, 'MC'],
-  [SpaceBonus.ANIMAL, 'Animal'],
-  [SpaceBonus.MICROBE, 'Microbe'],
-  [SpaceBonus.POWER, 'Power'],
-]);
-
-function bonusAsString(bonus: SpaceBonus | undefined): string {
-  return (bonus === undefined) ? 'undefined' : (bonusToStringMap.get(bonus)|| 'special');
-}
