+++ tests/ares/AresHandler.spec.ts
@@ -1,5 +1,4 @@
 import {expect} from 'chai';
-import {AresHandler} from '../../src/ares/AresHandler';
 import {SpaceBonus} from '../../src/SpaceBonus';
 import {Player} from '../../src/Player';
 import {Game} from '../../src/Game';
@@ -16,7 +15,11 @@ import {DesperateMeasures} from '../../src/cards/ares/DesperateMeasures';
 import {fail} from 'assert';
 import {Phase} from '../../src/Phase';
 import {TestPlayers} from '../TestingUtils';
+import {_AresHazardPlacement} from '../../src/ares/AresHazards';
+import {AresSetup} from '../../src/ares/AresSetup';
 
+// oddly, this no longer tests AresHandler calls. So that's interesting.
+// TODO(kberg): break up tests, but no rush.
 describe('AresHandler', function() {
   let player : Player; let otherPlayer: Player; let game : Game;
 
@@ -63,22 +66,22 @@ describe('AresHandler', function() {
       space.tile = undefined; space.player = undefined;
     });
 
-    AresHandler.setupHazards(game, 4);
+    AresSetup.setupHazards(game, 4);
 
-        interface SpaceToTest {
-            tile: ITile;
-            x: number;
-            y: number;
-        }
-        const spacesWithTiles: Array<SpaceToTest> = game.board.spaces
-          .filter((space) => space.tile !== undefined)
-          .map((space) => {
-            const x: SpaceToTest = {tile: space.tile!, x: space.x, y: space.y}; return x;
-          });
+    interface SpaceToTest {
+        tile: ITile;
+        x: number;
+        y: number;
+    }
+    const spacesWithTiles: Array<SpaceToTest> = game.board.spaces
+      .filter((space) => space.tile !== undefined)
+      .map((space) => {
+        const x: SpaceToTest = {tile: space.tile!, x: space.x, y: space.y}; return x;
+      });
 
-        expect(spacesWithTiles).to.deep.eq([
-          {tile: {tileType: TileType.DUST_STORM_MILD, protectedHazard: false}, x: 8, y: 0},
-          {tile: {tileType: TileType.DUST_STORM_MILD, protectedHazard: false}, x: 6, y: 8}]);
+    expect(spacesWithTiles).to.deep.eq([
+      {tile: {tileType: TileType.DUST_STORM_MILD, protectedHazard: false}, x: 8, y: 0},
+      {tile: {tileType: TileType.DUST_STORM_MILD, protectedHazard: false}, x: 6, y: 8}]);
   });
 
   it('Pay Adjacency Costs', function() {
@@ -119,7 +122,7 @@ describe('AresHandler', function() {
 
   it('Pay adjacent hazard costs - mild', function() {
     const firstSpace = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(firstSpace, TileType.DUST_STORM_MILD);
+    _AresHazardPlacement.putHazardAt(firstSpace, TileType.DUST_STORM_MILD);
 
     // No resources available to play the tile.
     player.addProduction(Resources.MEGACREDITS, -5);
@@ -142,7 +145,7 @@ describe('AresHandler', function() {
 
   it('pay adjacent hazard costs - severe', function() {
     const firstSpace = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(firstSpace, TileType.DUST_STORM_SEVERE);
+    _AresHazardPlacement.putHazardAt(firstSpace, TileType.DUST_STORM_SEVERE);
 
     // No resources available to play the tile.
     player.addProduction(Resources.MEGACREDITS, -5);
@@ -165,7 +168,7 @@ describe('AresHandler', function() {
 
   it('Adjacenct hazard costs do not apply to oceans', function() {
     const firstSpace = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(firstSpace, TileType.DUST_STORM_MILD);
+    _AresHazardPlacement.putHazardAt(firstSpace, TileType.DUST_STORM_MILD);
 
     const before = getProduction(player);
 
@@ -179,7 +182,7 @@ describe('AresHandler', function() {
 
   it('cover mild hazard', function() {
     const space = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(space, TileType.EROSION_MILD);
+    _AresHazardPlacement.putHazardAt(space, TileType.EROSION_MILD);
     player.megaCredits = 8;
     expect(player.getTerraformRating()).eq(20);
 
@@ -193,7 +196,7 @@ describe('AresHandler', function() {
 
   it('cover severe hazard', function() {
     const space = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(space, TileType.EROSION_SEVERE);
+    _AresHazardPlacement.putHazardAt(space, TileType.EROSION_SEVERE);
     player.megaCredits = 16;
     expect(player.getTerraformRating()).eq(20);
 
@@ -383,7 +386,7 @@ describe('AresHandler', function() {
 
   it('No adjacency hazard costs during WGT', function() {
     const firstSpace = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(firstSpace, TileType.DUST_STORM_MILD);
+    _AresHazardPlacement.putHazardAt(firstSpace, TileType.DUST_STORM_MILD);
     game.phase = Phase.SOLAR;
 
     const adjacentSpace = game.board.getAdjacentSpaces(firstSpace)[0];
@@ -395,7 +398,7 @@ describe('AresHandler', function() {
 
   it('No hazard coverage cost or bonus during WGT', function() {
     const space = game.board.getAvailableSpacesOnLand(player)[0];
-    AresHandler.putHazardAt(space, TileType.EROSION_SEVERE);
+    _AresHazardPlacement.putHazardAt(space, TileType.EROSION_SEVERE);
     player.megaCredits = 8;
     expect(player.getTerraformRating()).eq(20);
     game.phase = Phase.SOLAR;
